data_gen/step_1/data_processing/requirements.txt
python-dotenv
azure-storage-blob==12.2.0
spacy>=2.2.3
nltk
wikitextparser
mediawiki-parser
mwparserfromhell
tqdm
bs4
pywb
dateparser
langid
